Actually somewhat functioning storage system interaction in multiplayer.

+++ src/main/java/sunsetsatellite/retrostorage/util/INetworkController.java
@@ -14,7 +14,7 @@ import java.util.*;
 public interface INetworkController {
     boolean isActive();
 
-    int getEnergyConsumption();
+    long getEnergyConsumption();
 
     ArrayDeque<CraftingTask> getRequestQueue();
